Further corrections of typos and dependencies.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@259 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/conio.h
@@ -362,7 +362,7 @@ extern "C" {
  
 #endif
 
-  __CRT_INLINE unsigned __int64 __readmsr(int msr)
+  __CRT_INLINE unsigned __int64 __readmsr(unsigned long msr)
   {
       unsigned __int64 val1, val2;
        __asm__ __volatile__(
